requirements.txt
# CyberMind AI - Requirements
ollama>=0.3.0
customtkinter>=5.2.2
chromadb>=0.5.0
sentence-transformers>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiohttp>=3.9.0
pycryptodome>=3.20.0
pyperclip>=1.8.2
psutil>=5.9.0
packaging>=23.0
